Extract helper lemmas to utils and mark as library.
